+++ src/actions/fetchers/document/re-create-ai-pick/mutation.ts
@@ -3,7 +3,7 @@
 import { useMutation } from '@tanstack/react-query'
 import { useSession } from 'next-auth/react'
 import { reCreateAiPick } from '.'
-import useAmplitudeContext from '@/shared/hooks/use-amplitude-context'
+import { useAmplitudeContext } from '@/shared/hooks/use-amplitude-context'
 
 interface Params {
   documentId: number
