+++ src/app/(routes)/(tabs)/main/components/continuous-quiz-drawer-dialog.tsx
@@ -5,7 +5,7 @@ import Loading from '@/shared/components/loading'
 import { Dialog, DialogContent, DialogTrigger } from '@/shared/components/ui/dialog'
 import { Drawer, DrawerContent, DrawerTrigger } from '@/shared/components/ui/drawer'
 import icons from '@/constants/icons'
-import useAmplitudeContext from '@/shared/hooks/use-amplitude-context'
+import { useAmplitudeContext } from '@/shared/hooks/use-amplitude-context'
 import { useMediaQuery } from '@/shared/hooks/use-media-query'
 import { cn } from '@/shared/lib/utils'
 import Image from 'next/image'
