+++ src/app/(routes)/(tabs)/main/components/my-stars-drawer-dialog.tsx
@@ -3,7 +3,7 @@
 import { Dialog, DialogContent, DialogTrigger } from '@/shared/components/ui/dialog'
 import { Drawer, DrawerContent, DrawerTrigger } from '@/shared/components/ui/drawer'
 import icons from '@/constants/icons'
-import useAmplitudeContext from '@/shared/hooks/use-amplitude-context'
+import { useAmplitudeContext } from '@/shared/hooks/use-amplitude-context'
 import { useMediaQuery } from '@/shared/hooks/use-media-query'
 import { cn } from '@/shared/lib/utils'
 import Image from 'next/image'
