+++ src/app/(routes)/(tabs)/main/components/quiz-banner.tsx
@@ -13,9 +13,9 @@ import { CreateDocumentProtector } from '@/shared/components/create-document-pro
 import Loading from '@/shared/components/loading'
 import { SwitchCase } from '@/shared/components/react/switch-case'
 import { Button } from '@/shared/components/ui/button'
-import useAmplitudeContext from '@/shared/hooks/use-amplitude-context'
+import { useAmplitudeContext } from '@/shared/hooks/use-amplitude-context'
 import { cn } from '@/shared/lib/utils'
-import { calculateTimeUntilTomorrowMidnight, getCurrentDate } from '@/utils/date'
+import { calculateTimeUntilTomorrowMidnight, getCurrentDate } from '@/shared/utils/date'
 import { useQueryClient } from '@tanstack/react-query'
 import { useSession } from 'next-auth/react'
 import Image from 'next/image'
