+++ src/app/(routes)/(tabs)/review/components/quiz-analysis.tsx
@@ -5,7 +5,7 @@ import { useState } from 'react'
 import Loading from '@/shared/components/loading'
 import { QuizAnalysisSummary } from './ui/quiz-analysis-summary'
 import { QuizTypeChart } from './ui/quiz-type-chart'
-import { currentMonth } from '@/utils/date'
+import { currentMonth } from '@/shared/utils/date'
 import { Period } from './ui/period'
 import { PeriodTypeSelector } from './ui/period-type-selector'
 import { CategorySelect } from './ui/category-select'
