+++ src/app/(routes)/(tabs)/review/components/ui/period-type-selector.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import useAmplitudeContext from '@/shared/hooks/use-amplitude-context'
+import { useAmplitudeContext } from '@/shared/hooks/use-amplitude-context'
 import { cn } from '@/shared/lib/utils'
 
 interface Props {
