+++ src/app/(routes)/quiz/components/question.tsx
@@ -1,7 +1,7 @@
-import { getCurrentDate } from '@/utils/date'
+import { getCurrentDate } from '@/shared/utils/date'
 import ProgressBar from './progress-bar'
 import { motion } from 'framer-motion'
-import { msToElapsedTime } from '@/utils/time'
+import { msToElapsedTime } from '@/shared/utils/time'
 import { cn } from '@/shared/lib/utils'
 import { QuizType } from '@/actions/types/dto/quiz.dto'
