+++ src/app/(routes)/quiz/components/quiz-intro.tsx
@@ -1,5 +1,5 @@
 import { QuizDTO, QuizType } from '@/actions/types/dto/quiz.dto'
-import { getCurrentDate } from '@/utils/date'
+import { getCurrentDate } from '@/shared/utils/date'
 import { motion } from 'framer-motion'
 import { HTMLAttributes } from 'react'
