+++ src/app/(routes)/quiz/components/quiz.tsx
@@ -6,7 +6,7 @@ import { useEffect, useState } from 'react'
 import QuizHeader from './quiz-header'
 import Explanation from './explanation'
 import Question from './question'
-import { delay } from '@/utils/delay'
+import { delay } from '@/shared/utils/delay'
 import MultipleOptions from './multiple-options'
 import MixUpOptions from './mix-up-options'
 import { QuizProgress, SolvingData } from '../types'
@@ -22,8 +22,8 @@ import { ReportQuizError } from './report-quiz-error'
 import { deleteQuiz } from '@/actions/fetchers/quiz/delete-quiz'
 import { cn } from '@/shared/lib/utils'
 import { usePatchQuizResultMutation } from '@/actions/fetchers/quiz/patch-quiz-result/mutation'
-import useAmplitudeContext from '@/shared/hooks/use-amplitude-context'
-import { getCurrentTime } from '@/utils/date'
+import { useAmplitudeContext } from '@/shared/hooks/use-amplitude-context'
+import { getCurrentTime } from '@/shared/utils/date'
 
 interface QuizProps {
   quizzes: QuizDTO[]
