+++ src/app/(routes)/quiz/practice/components/example-quiz-header.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { Button } from '@/shared/components/ui/button'
-import { msToElapsedTime } from '@/utils/time'
+import { msToElapsedTime } from '@/shared/utils/time'
 import { useRouter } from 'next/navigation'
 import { HTMLAttributes } from 'react'
