+++ src/app/(routes)/quiz/practice/components/example-quiz-result.tsx
@@ -1,10 +1,10 @@
 'use client'
 
 import { Button } from '@/shared/components/ui/button'
-import { msToElapsedTime } from '@/utils/time'
+import { msToElapsedTime } from '@/shared/utils/time'
 import { motion } from 'framer-motion'
 import { useRouter } from 'next/navigation'
-import { getCurrentDate } from '@/utils/date'
+import { getCurrentDate } from '@/shared/utils/date'
 import { ExampleSolvingData } from './example-quiz'
 import { useMemo } from 'react'
