+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next'
 import './globals.css'
 import { cn } from '@/shared/lib/utils'
-import { dmSans, suit } from '@/lib/fonts'
+import { dmSans, suit } from '@/shared/lib/fonts'
 import { Metadatas } from '@/features/metadata'
 import { Providers } from '@/providers'
 import { Toaster } from '@/shared/components/ui/toaster'
