+++ src/providers/index.tsx
@@ -3,9 +3,8 @@
 import { ReactNode, useState } from 'react'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { getQueryClient } from '@/shared/lib/tanstack-query/client'
-import AmplitudeContextProvider from '@/contexts/amplitude-context'
 import { SessionProvider } from 'next-auth/react'
-import { ApiClientProvider } from '@/shared/lib/api-client-provider'
+import { AmplitudeContextProvider } from '@/shared/hooks/use-amplitude-context'
 
 export function Providers({ children }: { children: ReactNode }) {
   const [queryClient] = useState(() => getQueryClient())
@@ -13,9 +12,7 @@ export function Providers({ children }: { children: ReactNode }) {
   return (
     <AmplitudeContextProvider>
       <SessionProvider>
-        <ApiClientProvider>
-          <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-        </ApiClientProvider>
+        <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
       </SessionProvider>
     </AmplitudeContextProvider>
   )
