+++ src/shared/utils/assert.ts
